add fields in restart file
