+++ tests/test_restart.cpp
@@ -1,5 +1,6 @@
 #include <gtest/gtest.h>
 #include <samurai/box.hpp>
+#include <samurai/field.hpp>
 #include <samurai/io/restart.hpp>
 #include <samurai/mr/mesh.hpp>
 #include <samurai/uniform_mesh.hpp>
@@ -27,22 +28,22 @@ namespace samurai
         {
             auto mesh = LevelCellArray<1>(4, Box<double, 1>(0, 1));
             decltype(mesh) mesh2;
-            dump_mesh("mesh", mesh);
-            load_mesh("mesh", mesh2);
+            dump("mesh", mesh);
+            load("mesh", mesh2);
             EXPECT_TRUE(mesh == mesh2);
         }
         {
             auto mesh = LevelCellArray<2>(4, Box<double, 2>({0, 0}, {1, 1}));
             decltype(mesh) mesh2;
-            dump_mesh("mesh", mesh);
-            load_mesh("mesh", mesh2);
+            dump("mesh", mesh);
+            load("mesh", mesh2);
             EXPECT_TRUE(mesh == mesh2);
         }
         {
             auto mesh = LevelCellArray<3>(4, Box<double, 3>({0, 0, 0}, {1, 1, 1}));
             decltype(mesh) mesh2;
-            dump_mesh("mesh", mesh);
-            load_mesh("mesh", mesh2);
+            dump("mesh", mesh);
+            load("mesh", mesh2);
             EXPECT_TRUE(mesh == mesh2);
         }
     }
@@ -56,8 +57,8 @@ namespace samurai
             cl[4][{}].add_interval({0, 1});
             auto mesh = CellArray<1>(cl);
             decltype(mesh) mesh2;
-            dump_mesh("mesh", mesh);
-            load_mesh("mesh", mesh2);
+            dump("mesh", mesh);
+            load("mesh", mesh2);
             EXPECT_TRUE(mesh == mesh2);
         }
         {
@@ -67,8 +68,8 @@ namespace samurai
             cl[4][{6}].add_interval({0, 1});
             auto mesh = CellArray<2>(cl);
             decltype(mesh) mesh2;
-            dump_mesh("mesh", mesh);
-            load_mesh("mesh", mesh2);
+            dump("mesh", mesh);
+            load("mesh", mesh2);
             EXPECT_TRUE(mesh == mesh2);
         }
         {
@@ -78,8 +79,8 @@ namespace samurai
             cl[4][{6, 1}].add_interval({0, 1});
             auto mesh = CellArray<3>(cl);
             decltype(mesh) mesh2;
-            dump_mesh("mesh", mesh);
-            load_mesh("mesh", mesh2);
+            dump("mesh", mesh);
+            load("mesh", mesh2);
             EXPECT_TRUE(mesh == mesh2);
         }
     }
@@ -89,8 +90,8 @@ namespace samurai
         using Config = UniformConfig<2>;
         auto mesh    = UniformMesh<Config>(Box<double, 2>({0, 0}, {1, 1}), 3);
         decltype(mesh) mesh2;
-        dump_mesh("mesh", mesh);
-        load_mesh("mesh", mesh2);
+        dump("mesh", mesh);
+        load("mesh", mesh2);
         EXPECT_TRUE(mesh == mesh2);
     }
 
@@ -98,8 +99,37 @@ namespace samurai
     {
         auto mesh = create_mesh<2>(1);
         decltype(mesh) mesh2;
-        dump_mesh("mesh", mesh);
-        load_mesh("mesh", mesh2);
+        dump("mesh", mesh);
+        load("mesh", mesh2);
         EXPECT_TRUE(mesh == mesh2);
+    }
+
+    TEST(restart, restart_field)
+    {
+        auto mesh = create_mesh<2>(1);
+        auto u    = make_field<double, 1>("u", mesh);
+        u.fill(1.);
+        dump("mesh", mesh, u);
+
+        auto mesh2 = create_mesh<2>(10);
+        auto u2    = make_field<double, 1>("u", mesh2);
+        load("mesh", mesh2, u2);
+        EXPECT_TRUE(u == u2);
+    }
+
+    TEST(restart, restart_multiple_fields)
+    {
+        auto mesh = create_mesh<2>(1);
+        auto u    = make_field<double, 1>("u", mesh);
+        u.fill(1.);
+        auto v = make_field<int, 2>("v", mesh);
+        v.fill(2);
+        dump("mesh", mesh, u, v);
+        auto mesh2 = create_mesh<2>(10);
+        auto u2    = make_field<double, 1>("u", mesh2);
+        auto v2    = make_field<int, 2>("v", mesh2);
+        load("mesh", mesh2, u2, v2);
+        EXPECT_TRUE(u == u2);
+        EXPECT_TRUE(v == v2);
     }
 }
